refactored and committed spring beans
