+++ source/projects/demos/LoggingModifiers/META-INF/MANIFEST.MF
@@ -3,7 +3,7 @@ Bundle-ManifestVersion: 2
 Bundle-Name: Test
 Bundle-SymbolicName: LoggingModifiers
 Bundle-Version: 1.0.0.qualifier
-Bundle-Activator: edu.emory.cci.sample.qrm.bundle.Activator
+Bundle-Activator: edu.emory.cci.sample.modifiers.bundle.Activator
 Bundle-ActivationPolicy: lazy
 Bundle-RequiredExecutionEnvironment: JavaSE-1.6
 Import-Package: com.google.gson;version="2.1.0",
