+++ source/projects/demos/LoggingModifiers/META-INF/spring/beans.xml
@@ -1,0 +1,21 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<beans xmlns="http://www.springframework.org/schema/beans"
+	xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xmlns:osgi="http://www.springframework.org/schema/osgi"
+	xsi:schemaLocation="http://www.springframework.org/schema/beans   
+       http://www.springframework.org/schema/beans/spring-beans.xsd                      
+       http://www.springframework.org/schema/osgi  
+       http://www.springframework.org/schema/osgi/spring-osgi.xsd">
+	
+	<bean id="qm" class="edu.emory.cci.sample.modifiers.LoggingQM" >
+	</bean>
+
+	<bean id="qrm" class="edu.emory.cci.sample.modifiers.LoggingQRM">
+	</bean>
+
+	<bean id="spm" class="edu.emory.cci.sample.modifiers.LoggingSPM">
+	</bean>
+	
+</beans>
+
+
+
