+++ source/projects/demos/LoggingModifiers/src/edu/emory/cci/sample/modifiers/LoggingQRM.java
@@ -1,4 +1,4 @@
-package edu.emory.cci.sample.qrm;
+package edu.emory.cci.sample.modifiers;
 
 import java.io.ByteArrayInputStream;
 import java.io.ByteArrayOutputStream;
